pyproject.toml
[tool.poetry]
name = "noisetgbot"
version = "0.1.0"
description = ""
authors = ["neekee <Uspeshniy.ae@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.0rc1"
asyncpg = "^0.27.0"
sqlalchemy = "^2.0.12"
python-dotenv = "^1.0.0"
aiogram = "^2.25.1"
logging = "^0.4.9.6"
pydrive2 = "^1.15.3"
alembic = "^1.10.4"
python-dateutil = "^2.8.2"
google-api-core = "^2.11.0"
google-api-python-client = "^2.86.0"
google-auth = "^2.17.3"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
googleapis-common-protos = "^1.59.0"
oauth2client = "^4.1.3"
oauthlib = "^3.2.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
mypy = "^1.2.0"
pylint = "^2.17.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
